+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
  * Builtin functions to check for nan, infinity and subnormality in floats (#101)
  * Builtin random function (#102)
+ * Implement `TryFrom<Value>` for all types a value can hold (#105)
 
 ### Removed
 
@@ -22,6 +23,7 @@
 My warmhearted thanks goes to:
 
  * [Ophir LOJKINE](https://github.com/lovasoa)
+ * [Joe Grund](https://github.com/jgrund)
 
 ## [7.2.0](https://github.com/ISibboI/evalexpr/compare/7.1.1...7.2.0) - 2022-03-16